refactor($core): clarify palette.styl and style.styl
