+++ packages/@vuepress/core/lib/internal-plugins/style/client.js
@@ -1,0 +1,4 @@
+// generated from user config
+import('@temp/style.styl')
+
+window.__VUEPRESS_STYLE_STYLE_LOADED__ = true
