+++ packages/@vuepress/core/lib/prepare/AppContext.js
@@ -111,7 +111,8 @@ module.exports = class AppContext {
       .use(require('../internal-plugins/routes'))
       .use(require('../internal-plugins/rootMixins'))
       .use(require('../internal-plugins/enhanceApp'))
-      .use(require('../internal-plugins/overrideCSS'))
+      .use(require('../internal-plugins/palette'))
+      .use(require('../internal-plugins/style'))
       .use(require('../internal-plugins/layoutComponents'))
       .use(require('../internal-plugins/pageComponents'))
       .use(require('../internal-plugins/transformModule'))
